Extract and make available in component perform context
