+++ packages/spectral/src/serverTypes/convertIntegration.ts
@@ -27,7 +27,6 @@ import {
   TriggerReference,
   TriggerEventFunctionReturn,
   isOrganizationActivatedConnectionConfigVar,
-  DebugContext,
 } from "../types";
 import {
   Component as ServerComponent,
@@ -42,7 +41,6 @@ import {
   ActionContext,
   TriggerPayload,
   TriggerResult,
-  ActionPerformReturn,
 } from ".";
 import { convertInput } from "./convertComponent";
 import {
@@ -55,8 +53,7 @@ import {
 } from "./integration";
 import merge from "lodash/merge";
 import { createInvokeFlow } from "./perform";
-import { memoryUsage } from "node:process";
-import { performance } from "node:perf_hooks";
+import { createDebugContext } from "./context";
 
 export const convertIntegration = (definition: IntegrationDefinition): ServerComponent => {
   // Generate a unique reference key that will be used to reference the
@@ -954,60 +951,16 @@ const convertOnExecution =
       {},
     );
 
-    const globalDebug = Boolean(context.globalDebug);
-
-    const debugContext: DebugContext = {
-      enabled: globalDebug,
-      timeElapsed: {
-        mark: (actionContext: ActionContext, label: string) => {
-          if (globalDebug) {
-            actionContext.debug.results.timeElapsed.marks[label] = performance.now();
-          }
-        },
-        measure: (
-          actionContext: ActionContext,
-          label: string,
-          marks: { start: string; end: string },
-        ) => {
-          if (globalDebug) {
-            actionContext.debug.results.timeElapsed.measurements[label] = {
-              marks,
-              duration:
-                actionContext.debug.results.timeElapsed.marks[marks.end] -
-                actionContext.debug.results.timeElapsed.marks[marks.start],
-            };
-          }
-        },
-      },
-      memoryUsage: (actionContext: ActionContext, label: string, showDetail?: boolean) => {
-        if (globalDebug) {
-          // @ts-expect-error: memoryUsage.rss() is documented but not typed
-          const usage = showDetail ? memoryUsage() : (memoryUsage.rss() as number);
-
-          actionContext.debug.results.memoryUsage.push({
-            mark: label,
-            rss: typeof usage === "number" ? usage / 1000000 : usage.rss / 1000000,
-            detail: typeof usage === "number" ? undefined : usage,
-          });
-        }
-      },
-      results: {
-        timeElapsed: { marks: {}, measurements: {} },
-        memoryUsage: [],
-        allowedMemory: Number(context.runnerAllocatedMemoryMb),
-      },
-    };
-
     const actionContext = {
       ...context,
-      debug: debugContext,
+      debug: createDebugContext(context),
       components: componentMethods,
       invokeFlow: createInvokeFlow(context, { isCNI: true }),
     };
 
     const result = await onExecution(actionContext, params);
 
-    if (globalDebug) {
+    if (context.globalDebug) {
       context.logger.metric(actionContext.debug.results);
     }
